chore: move content insights location
