+++ src/apps/content-editor/src/app/views/ItemEdit/Meta/ContentInsights/WordCount.tsx
